disable auto reboot on cli

+++ archinstall.sh
@@ -87,7 +87,8 @@ sudo pkill instantmenu
 sleep 2
 
 # ask to reboot, upload error data if install failed
-if ! [ -e /opt/installfailed ] || ! [ -e /opt/installsuccess ]; then
+if ! [ -e /opt/installfailed ] || ! [ -e /opt/installsuccess ] &&
+    command -v installapplet; then
     notify-send "rebooting"
     sleep 2
     reboot
